requirements.txt
bcrypt==4.3.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==45.0.7
idna==3.10
invoke==2.2.0
paramiko==4.0.0
pycparser==2.22
PyNaCl==1.5.0
PyYAML==6.0.2
requests==2.32.5
urllib3==2.5.00
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==45.0.7
idna==3.10
invoke==2.2.0
paramiko==4.0.0
pycparser==2.22
PyNaCl==1.5.0
requests==2.32.5
urllib3==2.5.0
